Moved files to reflect logic
